transfer learning on PASCAL dataset/transfer_learning/requirements.txt
absl-py==0.1.10
appnope==0.1.0
astor==0.6.2
backports-abc==0.5
beautifulsoup4==4.3.2
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
CVXcanon==0.1.1
cvxpy==0.4.11
cycler==0.10.0
decorator==4.2.1
dill==0.2.7.1
ecos==2.0.5
entrypoints==0.2.3
fastcache==1.0.2
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
more-itertools==4.1.0
multiprocess==0.70.5
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
scs==2.0.2
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.1
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
voc-utils==0.0
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.1.4
